+++ app/src/main/AndroidManifest.xml
@@ -14,10 +14,13 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31">
         <activity
-            android:name=".ProductDetailActivity"
+            android:name=".shoppingcart.ShoppingCartActivity"
             android:exported="false" />
         <activity
-            android:name=".ShoppingActivity"
+            android:name=".productdetail.ProductDetailActivity"
+            android:exported="false" />
+        <activity
+            android:name=".shopping.ShoppingActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
