+++ app/src/main/java/woowacourse/shopping/shopping/ShoppingActivity.kt
@@ -1,9 +1,12 @@
-package woowacourse.shopping
+package woowacourse.shopping.shopping
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.R
+import woowacourse.shopping.productdetail.ProductDetailActivity
+import woowacourse.shopping.productdetail.ProductUiModel
 
 class ShoppingActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
