+++ app/src/main/java/woowacourse/shopping/shopping/ShoppingItemViewHolder.kt
@@ -1,10 +1,11 @@
-package woowacourse.shopping
+package woowacourse.shopping.shopping
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import woowacourse.shopping.databinding.ItemShoppingProductBinding
+import woowacourse.shopping.productdetail.ProductUiModel
 
 class ShoppingItemViewHolder(
     private val binding: ItemShoppingProductBinding
