+++ app/src/main/res/layout/activity_product_detail.xml
@@ -1,40 +1,40 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ProductDetailActivity">
+        tools:context=".productdetail.ProductDetailActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
-            android:background="#555555"
             android:id="@+id/product_detail_tool_bar"
             android:layout_width="match_parent"
             android:layout_height="56dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent">
+            android:background="#555555"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent">
 
             <ImageView
                 android:id="@+id/image_cancel"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:src="@drawable/cancel_24"
                 android:layout_marginEnd="25dp"
-                app:layout_constraintEnd_toEndOf="parent"
+                android:src="@drawable/cancel_24"
                 app:layout_constraintBottom_toBottomOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintTop_toTopOf="parent" />
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <LinearLayout
             android:layout_width="0dp"
             android:layout_height="0dp"
-            app:layout_constraintTop_toBottomOf="@id/product_detail_tool_bar"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintEnd_toEndOf="parent"
+            android:orientation="vertical"
             app:layout_constraintBottom_toBottomOf="parent"
-            android:orientation="vertical">
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/product_detail_tool_bar">
 
             <ImageView
                 android:id="@+id/image_product_detail"
@@ -46,20 +46,20 @@
                 android:id="@+id/text_product_name"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                tools:text="아메리카노"
-                android:textSize="24sp"
-                android:layout_marginTop="16dp"
                 android:layout_marginStart="18dp"
-                android:textColor="@color/black" />
+                android:layout_marginTop="16dp"
+                android:textColor="@color/black"
+                android:textSize="24sp"
+                tools:text="아메리카노" />
 
             <TextView
                 android:id="@+id/text_product_price"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                tools:text="금액: 99800"
-                android:textSize="20sp"
+                android:layout_marginStart="18dp"
                 android:layout_marginTop="29dp"
-                android:layout_marginStart="18dp" />
+                android:textSize="20sp"
+                tools:text="금액: 99800" />
 
 
         </LinearLayout>
@@ -68,11 +68,11 @@
             android:id="@+id/button_put_to_shopping_cart"
             android:layout_width="match_parent"
             android:layout_height="50dp"
-            app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            android:textSize="20sp"
+            android:background="#04C09E"
             android:text="@string/put_to_shopping_cart"
-            android:background="#04C09E" />
+            android:textSize="20sp"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintStart_toStartOf="parent" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
