+++ app/src/main/res/layout/item_shopping_product.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="wrap_content"
@@ -9,33 +9,33 @@
         android:padding="10dp">
 
         <ImageView
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
             android:id="@+id/image_product"
             android:layout_width="154dp"
             android:layout_height="154dp"
-            android:src="@drawable/ic_launcher_background" />
+            android:src="@drawable/ic_launcher_background"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
 
         <TextView
             android:id="@+id/text_product_name"
             android:layout_width="140dp"
             android:layout_height="wrap_content"
-            tools:text="PET보틀-정사각..."
-            android:textSize="16sp"
-            android:textColor="#333333"
             android:layout_marginTop="10dp"
+            android:textColor="#333333"
+            android:textSize="16sp"
+            app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/image_product"
-            app:layout_constraintEnd_toEndOf="parent" />
+            tools:text="PET보틀-정사각..." />
 
         <TextView
             android:id="@+id/text_product_price"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            app:layout_constraintTop_toBottomOf="@id/text_product_name"
             android:textSize="16sp"
-            tools:text="10,000원"
-            app:layout_constraintStart_toStartOf="@+id/text_product_name" />
+            app:layout_constraintStart_toStartOf="@+id/text_product_name"
+            app:layout_constraintTop_toBottomOf="@id/text_product_name"
+            tools:text="10,000원" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
