+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,5 @@
 <resources>
     <string name="app_name">Shopping</string>
     <string name="put_to_shopping_cart">장바구니 담기</string>
+    <string name="cart">Cart</string>
 </resources>
